don't use int64 in depjobs for now

+++ models/models.go
@@ -145,7 +145,7 @@ type BatchJob struct {
 }
 
 type DepJob struct {
-	ID     int64         `gorm:"primary_key" json:"-"`
+	ID     int           `gorm:"primary_key" json:"-"`
 	DepId  string        `json:"-" validate:"nonzero"`
 	Events []DepJobEvent `json:"events" gorm:"ForeignKey:BatchJobId"`
 }
@@ -194,9 +194,8 @@ func (d *DepJob) HasFinished() bool {
 }
 
 type DepJobEvent struct {
-
-	ID        int64     `gorm:"primary_key" json:"-"`
-	DepJobId  int64     `json:"-" validate:"nonzero"`
+	ID        int       `gorm:"primary_key" json:"-"`
+	DepJobId  int       `json:"-" validate:"nonzero"`
 	Timestamp time.Time `json:"timestamp"`
 	Status    string    `json:"status"`
 	Message   string    `json:"message,omitempty"`
